refactor pattern & lib
