+++ lib/components/note_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_week2/note.dart';
-import 'constants.dart';
+import 'package:flutter_week2/model/note.dart';
+import '../constants.dart';
 import 'package:intl/intl.dart';
