+++ lib/model/note.dart
